requirements.txt
# YouTube API
google-api-python-client==2.108.0

# Environment and configuration
python-dotenv==1.0.0

# Kafka
kafka-python==2.0.2

# Spark Processing
pyspark==3.5.0

# Azure Cloud
azure-identity==1.15.0
azure-storage-file-datalake==12.12.0

# Data Processing
pandas==2.1.3
numpy==1.26.0

# Monitoring and Scheduling
apscheduler==3.10.4
tabulate==0.9.0

# Development and Testing
black==23.11.0
flake8==6.1.0
pytest==7.4.3

# Utilities
pyarrow==14.0.1  # For parquet file handling